refactor: more abstraction of RecipeLogics
