+++ src/main/kotlin/com/github/trc/clayium/api/capability/impl/AbstractRecipeLogic.kt
@@ -6,6 +6,7 @@ import com.cleanroommc.modularui.value.sync.GuiSyncManager
 import com.cleanroommc.modularui.value.sync.SyncHandlers
 import com.cleanroommc.modularui.widgets.ProgressWidget
 import com.github.trc.clayium.api.ClayEnergy
+import com.github.trc.clayium.api.capability.AbstractWorkable
 import com.github.trc.clayium.api.capability.ClayiumDataCodecs
 import com.github.trc.clayium.api.capability.IControllable
 import com.github.trc.clayium.api.metatileentity.MTETrait
@@ -20,31 +21,20 @@ import com.github.trc.clayium.integration.jei.JeiPlugin
 import net.minecraft.item.ItemStack
 import net.minecraft.nbt.NBTTagCompound
 
+/**
+ * Recipe-based implementation of [AbstractWorkable]
+ */
 abstract class AbstractRecipeLogic(
     metaTileEntity: MetaTileEntity,
     val recipeRegistry: RecipeRegistry<*>,
-) : MTETrait(metaTileEntity, ClayiumDataCodecs.RECIPE_LOGIC), IControllable {
+) : AbstractWorkable(metaTileEntity), IControllable {
 
     protected val inputInventory = metaTileEntity.importItems
 
     protected var previousRecipe: Recipe? = null
     protected var recipeCEt = ClayEnergy.ZERO
-    var requiredProgress = 0L
-        protected set
-    var currentProgress = 0L
-        protected set
     // item stacks that will be produced when the recipe is done
     protected var itemOutputs: List<ItemStack> = emptyList()
-
-    protected var invalidInputsForRecipes = false
-    protected var outputsFull = false
-
-    override val isWorking: Boolean get() = currentProgress != 0L
-    override var isWorkingEnabled: Boolean = true
-        set(value) {
-            field = value
-            metaTileEntity.markDirty()
-        }
 
     /**
      * Draw energy from the energy container
@@ -54,50 +44,23 @@ abstract class AbstractRecipeLogic(
      */
     protected abstract fun drawEnergy(ce: ClayEnergy, simulate: Boolean = false): Boolean
 
-    override fun update() {
-        if (metaTileEntity.world?.isRemote == true) return
-        if (!isWorkingEnabled) return
-        if (currentProgress != 0L) {
-            updateRecipeProgress()
-        }
-        if (currentProgress == 0L && shouldSearchNewRecipe()) {
-            trySearchNewRecipe()
-        }
+    override fun showRecipesInJei() {
+        JeiPlugin.jeiRuntime.recipesGui.showCategories(listOf(this.recipeRegistry.category.uniqueId))
     }
 
-    protected open fun updateRecipeProgress() {
+    override fun updateWorkingProgress() {
         if (drawEnergy(recipeCEt)) currentProgress++
         if (currentProgress > requiredProgress) {
-            completeRecipe()
+            completeWork()
         }
     }
 
-    protected open fun completeRecipe() {
+    override fun completeWork() {
         currentProgress = 0
         TransferUtils.insertToHandler(metaTileEntity.exportItems, itemOutputs)
     }
 
-    private fun shouldSearchNewRecipe(): Boolean {
-        return canWorkWithInputs() && canFitNewOutputs()
-    }
-
-    private fun canWorkWithInputs(): Boolean {
-        if (invalidInputsForRecipes && !metaTileEntity.hasNotifiedInputs) return false
-
-        invalidInputsForRecipes = false
-        metaTileEntity.hasNotifiedInputs = false
-        return true
-    }
-
-    private fun canFitNewOutputs(): Boolean {
-        if (outputsFull && !metaTileEntity.hasNotifiedOutputs) return false
-
-        outputsFull = false
-        metaTileEntity.hasNotifiedOutputs = false
-        return true
-    }
-
-    protected open fun trySearchNewRecipe() {
+    override fun trySearchNewRecipe() {
         var currentRecipe: Recipe? = null
         currentRecipe = if (previousRecipe?.matches(false, inputInventory, tierNum) == true) {
             previousRecipe
@@ -127,49 +90,15 @@ abstract class AbstractRecipeLogic(
         this.previousRecipe = recipe
     }
 
-    fun getProgressBar(syncManager: GuiSyncManager): ProgressWidget {
-        syncManager.syncValue("requiredProgress", SyncHandlers.longNumber(
-            { requiredProgress },
-            { rProgress -> requiredProgress = rProgress }
-        ))
-        syncManager.syncValue("craftingProgress", SyncHandlers.longNumber(
-            { currentProgress },
-            { cProgress -> currentProgress = cProgress }
-        ))
-
-        val widget = ProgressWidget()
-            .size(22, 17)
-            .progress(this::getNormalizedProgress)
-            .texture(ClayGuiTextures.PROGRESS_BAR, 22)
-            if (Mods.JustEnoughItems.isModLoaded) {
-                widget.addTooltipLine(IKey.lang("jei.tooltip.show.recipes"))
-                    .listenGuiAction(IGuiAction.MousePressed { _ ->
-                        if (!widget.isBelowMouse) return@MousePressed false
-                        JeiPlugin.jeiRuntime.recipesGui.showCategories(listOf(this@AbstractRecipeLogic.recipeRegistry.category.uniqueId))
-                        return@MousePressed true
-                })
-            }
-
-        return widget
-    }
-
-    fun getNormalizedProgress(): Double {
-        if (currentProgress == 0L || requiredProgress == 0L) return 0.0
-        return (currentProgress.toDouble() - 1.0) / requiredProgress.toDouble()
-    }
-
     override fun serializeNBT(): NBTTagCompound {
         val data = super.serializeNBT()
-        data.setLong("currentProgress", currentProgress)
-        data.setLong("requiredProgress", requiredProgress)
         CUtils.writeItems(itemOutputs, "itemOutputs", data)
         data.setLong("recipeCEt", recipeCEt.energy)
         return data
     }
 
     override fun deserializeNBT(data: NBTTagCompound) {
-        currentProgress = data.getLong("currentProgress")
-        requiredProgress = data.getLong("requiredProgress")
+        super.deserializeNBT(data)
         itemOutputs = CUtils.readItems("itemOutputs", data)
         recipeCEt = ClayEnergy(data.getLong("recipeCEt"))
     }
